feat: transactions-issues
